# openapi/openapi.yaml
{
  "openapi": "3.0.4",
  "info": {
    "title": "Equipment Service Desk",
    "version": "0.1.0",
    "description": "Original synthetic single-organization service desk. Local demonstration only; no production deployment claim.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8089",
      "description": "Loopback-only synthetic development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/service-requests": {
      "get": {
        "operationId": "requestList",
        "summary": "Role-scoped service requests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "submitted",
                "triaged",
                "assigned",
                "resolved",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestPage"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request or infrastructure failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "requestSubmit",
        "summary": "Requester submits their own request",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequest"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request or infrastructure failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests/{id}": {
      "get": {
        "operationId": "requestGet",
        "summary": "Role-scoped request and current version",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequest"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request or infrastructure failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/service-requests/{id}/history": {
      "get": {
        "operationId": "requestHistory",
        "summary": "Role-scoped application audit history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEntryPage"
                }
              }
            },
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Rejected request or infrastructure failure",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Locally provisioned, short-lived opaque token. Authorization header only."
      }
    },
    "schemas": {
...
      "Triage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "policyVersion",
          "criticality",
          "impact",
          "priority",
          "targetSeconds",
          "dueAt"
        ],
        "properties": {
          "policyVersion": {
            "type": "string"
          },
          "criticality": {
            "type": "string"
          },
          "impact": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "targetSeconds": {
            "type": "integer"
          },
          "dueAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "nullable": true
      },
      "ServiceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "equipmentId",
          "requesterId",
          "title",
          "description",
          "reportedImpact",
          "state",
          "version",
          "submittedAt",
          "triage",
          "technicianId",
          "resolution",
          "resolvedAt",
          "cancellationReason",
          "cancelledAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "equipmentId": {
            "type": "string",
            "format": "uuid"
          },
          "requesterId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "reportedImpact": {
            "type": "string",
            "enum": [
              "stopped",
              "degraded"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "submitted",
              "triaged",
              "assigned",
              "resolved",
              "cancelled"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "triage": {
            "$ref": "#/components/schemas/Triage"
          },
          "technicianId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "resolution": {
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "AuditEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "actorId",
          "action",
          "details",
          "occurredAt",
          "correlationId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "actorId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "action": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "correlationId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          },
          "violations": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "field",
                "code"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "SubmitRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "equipmentId",
          "title",
          "description",
          "reportedImpact"
        ],
        "properties": {
          "equipmentId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          },
          "reportedImpact": {
            "type": "string",
            "enum": [
              "stopped",
              "degraded"
            ]
          }
        }
      },
...
      "ServiceRequestPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page",
          "limit"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceRequest"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
      "AuditEntryPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page",
          "limit"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            }
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      }
    }
  }
}